% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{cancel_order}
\alias{cancel_order}
\title{Function to cancel an exiting order.}
\usage{
cancel_order(object, variety, order_id, parent_order_id = NULL)
}
\arguments{
\item{object}{An object of type kite connect with valid api_key and
access_token.}

\item{variety}{Order variety, e.g. regular, bo, co or amo.}

\item{order_id}{Order ID for the order to modify.}

\item{parent_order_id}{Required for bracket orders (cannot be modified).}
}
\value{
Returns an order ID (string), if successful.
}
\description{
Function to cancel an exiting order with specified ID.
}
\details{
This places a request to cancel an existing order, if it is not
already executed. Please note: no error checks are done for the input
parameters. Sanity check is left to the user of this function. If
successful, this will return the order ID (should be same as the
one passed). A successfully placed request does not mean it is
a valid order, and orders with erroneous user input can be immedeately
cancelled. Therefore it is good practice to check the order details once
an order ID is received from this call.
}
\seealso{
\code{\link[kiteconnect3]{order_history}}

\code{\link[kiteconnect3]{place_order}}

\code{\link[kiteconnect3]{modify_order}}
}
