% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{fetch_access_token}
\alias{fetch_access_token}
\title{Fetch access token given the api key, secret and a request token.}
\usage{
fetch_access_token(object)
}
\arguments{
\item{object}{An object of type kite connect, must have request token set.}
}
\value{
A string containing the access token.
}
\description{
This function returns an access token given the inputs, that
can be used for subsequent api calls.
}
\details{
This function generate the `access_token` by exchanging
`request_token`.The `request_token` obtained after the login flow. the
`access_token` required for all subsequent requests. The object passed in
this call must already have `request_token` set (along with api_key and
api_secret). A successful call also set the user data within the object.
}
\seealso{
\code{\link[kiteconnect3]{set_request_token}}
}
