% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{get_historical_price}
\alias{get_historical_price}
\title{Function to obtain current historical prices for a instrument.}
\usage{
get_historical_price(object, instrument_token, from, to, interval,
  continuous = 0)
}
\arguments{
\item{object}{An object of type kite connect with valid api_key and
access_token.}

\item{instrument_token}{An instrument_token. See details below.}

\item{from}{Starting time for history.}

\item{to}{Ending time for history.}

\item{interval}{Interval definition.}

\item{continuous}{True/ False for fetching continuous data (e.g. futures).}
}
\value{
an xts object with OHLC data.
}
\description{
Function to obtain historical prices with given interval.
}
\details{
This function returns OHLC price history of a given instrument.
The instrument token is the one obtained with a call to get_instruments.
The valid intervals are `minute`,`day`,`3minute`,`5minute`,`10minute`,
`15minute`,`30minute` and `60minute`. The parameters `from` and `to` should
be in YYYY-MM-DD HH:MM:SS format (no timezone adjustment is applied).
Please note: this function is meant to be called as a one-off way to get
the quotes.One should not call this function too many times. The api
possibly has a rate limit and too many requests will block the user. For
continous data, websocket is more appropriate choice. Check the Kite Connect online
documentation for more details.
}
\seealso{
\code{\link[kiteconnect3]{get_instruments}}

\code{\link[kiteconnect3]{get_quote}}

\code{\link[kiteconnect3]{get_ohlc}}

\code{\link[kiteconnect3]{get_ltp}}
}

