% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{get_holdings}
\alias{get_holdings}
\title{Function to get the holdings details.}
\usage{
get_holdings(object)
}
\arguments{
\item{object}{An object of type kite connect with valid api_key and
access_token.}
}
\value{
Returns a list with holdings details, if successful.
}
\description{
Gets the holdings details of the user for the day.
}
