% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{get_instruments}
\alias{get_instruments}
\title{Function to obtain list of all instruments.}
\usage{
get_instruments(object, exchange = NULL)
}
\arguments{
\item{object}{An object of type kite connect with valid api_key and
access_token.}

\item{exchange}{Exchange for the instruments list. If not specified, returns
list of instruments across all exchanges.}
}
\value{
Returns a dataframe of the instrument lists.
}
\description{
Function to obtain a list of tradable instruments.
}
\details{
This function returns list of instruments trading on a given
exchange or across all exchanges. Ideally one should not be required to
call this function more than once a day (and save a local copy).
}
