% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{get_login_url}
\alias{get_login_url}
\title{Function to get the login URL.}
\usage{
get_login_url(object)
}
\arguments{
\item{object}{An object of kite connect type.}
}
\value{
Returns the login URL.
}
\description{
Function to get the login URL (the login entry + api_key) from
the kite object.
}
