% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{get_margins}
\alias{get_margins}
\title{Function to get the details of margins for the logged on user.}
\usage{
get_margins(object, segment = NULL)
}
\arguments{
\item{object}{An object of type kite connect with valid api_key and
access_token.}

\item{segment}{Either `equity` or `commodity`.}
}
\value{
Returns a list with margin details, if successful.
}
\description{
Gets the margin details of the user.
}
\details{
Returns a list with margin details. If no segment is specified.
returns for both the segments, else only for the one specified.
}
