% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{get_ohlc}
\alias{get_ohlc}
\title{Function to obtain current OHLC prices for a list of instruments.}
\usage{
get_ohlc(object, instruments)
}
\arguments{
\item{object}{An object of type kite connect with valid api_key and
access_token.}

\item{instruments}{A list of instruments. See details below.}
}
\value{
a dataframe with OHLC data, with instruments as row names.
}
\description{
Function to obtain open/high/low/close prices.
}
\details{
This function accepts a list of instruments (as exchange:symbol
pair, e.g. NFO:NIFTY18JANFUT) and returns current OHLC prices. Please note:
this function is meant to be called as a one-off way to get the quotes.
One should not call this function too many times. The api possibly has a
rate limit and too many requests will block the user. For continous data,
websocket is more appropriate choice. Check the Kite Connect online
documentation for more details.
}
\seealso{
\code{\link[kiteconnect3]{get_quote}}

\code{\link[kiteconnect3]{get_ltp}}

\code{\link[kiteconnect3]{get_historical_price}}
}

