% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{get_positions}
\alias{get_positions}
\title{Function to get the position details.}
\usage{
get_positions(object)
}
\arguments{
\item{object}{An object of type kite connect with valid api_key and
access_token.}
}
\value{
Returns a list of two dataframe with trade details, if successful,
one for net (`net`) positions and another for the day(`day`).
}
\description{
Gets the position details of the user for the day and net.
}
