% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{get_profile}
\alias{get_profile}
\title{Function to get the details of logged on user.}
\usage{
get_profile(object)
}
\arguments{
\item{object}{An object of type kite connect with valid api_key and
access_token.}
}
\value{
Returns a list with user details, if successful.
}
\description{
Gets the details of the user.
}
