% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{get_quote}
\alias{get_quote}
\title{Function to obtain current quote for a list of instruments.}
\usage{
get_quote(object, instruments)
}
\arguments{
\item{object}{An object of type kite connect with valid api_key and
access_token.}

\item{instruments}{A list of instruments. See details below.}
}
\value{
a list of quote and market depth data, with row names as
instruments.
}
\description{
Function to obtain quotes and market depths.
}
\details{
This function accepts a list of instruments (as exchange:symbol
pair, e.g. NFO:NIFTY18JANFUT) and returns current market quotes, including
last traded price, OHLC prices and market depth (up to 5 best bids and
offers). Please note: this function is meant to be called as a one-off
way to get the quotes. One should not call this function too many times.
The api possibly has a rate limit and too many requests will block the
user. For continous data, websocket is more appropriate choice. Check the
Kite Connect online documentation for more details.
}
\seealso{
\code{\link[kiteconnect3]{get_ohlc}}

\code{\link[kiteconnect3]{get_ltp}}

\code{\link[kiteconnect3]{get_historical_price}}
}

