% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{get_trades}
\alias{get_trades}
\title{Function to get the trade details.}
\usage{
get_trades(object)
}
\arguments{
\item{object}{An object of type kite connect with valid api_key and
access_token.}
}
\value{
Returns a dataframe with trade details, if successful.
}
\description{
Gets the order details of the user for the day.
}

