% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{kill_connection}
\alias{kill_connection}
\title{Function to end an api session.}
\usage{
kill_connection(object)
}
\arguments{
\item{object}{of type kite connect, must have a valid access token set.}
}
\value{
An object of type kite connect, with tokens reset.
}
\description{
This function invalidates and rest the access token.
}
\details{
This function sends a request to kite api to invalidate the
session and then reset the request token and access token within the object.
}

