% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{modify_order}
\alias{modify_order}
\title{Function to modify an exiting order.}
\usage{
modify_order(object, variety, order_id, parent_order_id = NULL,
  quantity = NULL, price = NULL, order_type = NULL,
  trigger_price = NULL, validity = NULL, disclosed_quantity = NULL)
}
\arguments{
\item{object}{An object of type kite connect with valid api_key and
access_token.}

\item{variety}{Order variety, e.g. regular, bo, co or amo.}

\item{order_id}{Order ID for the order to modify.}

\item{parent_order_id}{Required for bracket orders (cannot be modified).}

\item{quantity}{New quantity.}

\item{price}{New Price (e.g. for limit orders).}

\item{order_type}{Order type, e.g. MARKET, LIMIT, SL or SL-M.}

\item{trigger_price}{Trigger price if required (e.g. for stop loss orders).}

\item{validity}{Validity period of the order (DAY/IOC).}

\item{disclosed_quantity}{Quantity to disclosed publicly.}
}
\value{
Returns an order ID (string), if successful.
}
\description{
Function to modify an exiting order with specified details.
}
\details{
This function sends a modification request for an existing order,
assuming it is not already executed. Please note: not all parameters are
relevant for all types of orders. No error checks is carried out, It is
left to the function calling this method to carry out necessary error
checks.If successful, this will return the order ID (should be same as the
one passed). A successfully placed request does not mean it is
a valid order, and orders with erroneous user input can be immedeately
cancelled. Therefore it is good practice to check the order details once
an order ID is received from this call.
}
\seealso{
\code{\link[kiteconnect3]{order_history}}

\code{\link[kiteconnect3]{place_order}}

\code{\link[kiteconnect3]{cancel_order}}
}

