% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{order_history}
\alias{order_history}
\title{Function to get the order history.}
\usage{
order_history(object, order_id)
}
\arguments{
\item{object}{An object of type kite connect with valid api_key and
access_token.}

\item{order_id}{ID of the order to fetch the history for.}
}
\value{
Returns a dataframe with order history, if successful.
}
\description{
Gets the order history for a given order.
}
\seealso{
\code{\link[kiteconnect3]{order_trades}}
}
