% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{order_trades}
\alias{order_trades}
\title{Function to get the trade history.}
\usage{
order_trades(object, order_id)
}
\arguments{
\item{object}{An object of type kite connect with valid api_key and
access_token.}

\item{order_id}{ID of the order to fetch the history for.}
}
\value{
Returns a dataframe with trade history, if successful.
}
\description{
Gets the trade history for a given order.
}
\seealso{
\code{\link[kiteconnect3]{order_history}}
}
