% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{place_order}
\alias{place_order}
\title{Function to place a trade.}
\usage{
place_order(object, exchange, tradingsymbol, transaction_type, quantity,
  order_type = ORDER_TYPE_MARKET, variety = VARIETY_REGULAR,
  product = PRODUCT_NRML, price = NULL, validity = NULL,
  disclosed_quantity = NULL, trigger_price = NULL, squareoff = NULL,
  stoploss = NULL, trailing_stoploss = NULL, tag = NULL)
}
\arguments{
\item{object}{An object of type kite connect with valid api_key and
access_token.}

\item{exchange}{Exchange for the order, e.g. NSE, NFO, CDS etc.}

\item{tradingsymbol}{Trading Symbol of the instrument.}

\item{transaction_type}{BUY or SELL.}

\item{quantity}{Quantity of the order.}

\item{order_type}{Order type, e.g. MARKET, LIMIT, SL or SL-M.}

\item{variety}{Order variety, e.g. regular, bo, co or amo.}

\item{product}{Product type, e.g. MIS, NRML, CNC, BO or CO.}

\item{price}{Price if required (e.g. for limit orders).}

\item{validity}{Validity period of the order (DAY/IOC), defaults to DAY.}

\item{disclosed_quantity}{Quantity to disclosed publicly.}

\item{trigger_price}{Trigger price if required (e.g. for stop loss orders).}

\item{squareoff}{Price difference for profit booking (bracker orders).}

\item{stoploss}{Price difference for loss booking (bracket orders).}

\item{trailing_stoploss}{Trailing stop loss (for bracket orders).}

\item{tag}{Optional string to tag an order (alphanumeric, max 8 chars).}
}
\value{
Returns an order ID (string), if successful.
}
\description{
Function to place a trade with specified details.
}
\details{
This function sends an order with input details for execution.
Please note not all parameters are relevant for all types of orders. This
function (as with rest of the package) does no error checks and post the
trade request as is. It is left to the function calling this method to
carry out necessary error checks. If the order is successfully posted, an
order ID will be returned. A successfully posted order does not mean it is
a valid order, and orders with erroneous user input can be immedeately
cancelled. Therefore it is good practice to check the order details once
an order ID is received from this call.
}
\seealso{
\code{\link[kiteconnect3]{order_history}}

\code{\link[kiteconnect3]{modify_order}}

\code{\link[kiteconnect3]{cancel_order}}
}
